backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
